Add call login and logout

+++ catalog/application.py
@@ -86,33 +86,63 @@ def gconnect():
     stored_gplus_id = login_session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
         response = make_response(json.dumps('Current user is already' +
-                                 'connected.'), 200)
+                                            'connected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
-
     # Store the access token in the session for later use.
     login_session['access_token'] = credentials.access_token
     login_session['gplus_id'] = gplus_id
-
     # Get user info
     userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
     params = {'access_token': credentials.access_token, 'alt': 'json'}
     answer = requests.get(userinfo_url, params=params)
-
     data = answer.json()
-
     login_session['username'] = data['name']
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
-
     output = ''
     output += '<h1>Welcome, '
     output += login_session['username']
     output += '!</h1>'
     output += '<img src="'
     output += login_session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    output += ' " style = "width: 300px; height: 300px;border-radius:150px;'
+    output += '-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
     return output
+
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        print 'Access Token is None'
+        response = make_response(json.dumps('Current user not connected.'),
+                                 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    print 'In gdisconnect access token is %s', access_token
+    print 'User name is: '
+    print login_session['username']
+    url = 'https://accounts.google.com/o/oauth2'
+    '/revoke?token=%s' % login_session['access_token']
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'result is '
+    print result
+    if result['status'] == '200':
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        response = make_response(json.dumps('Failed to revoke token for given'
+                                            'user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
 
 
 @app.route('/category/<string:category_name>/json')
@@ -135,6 +165,8 @@ def menu(category_name):
 
 @app.route('/categoryItem/<string:category_name>/new', methods=['GET', 'POST'])
 def newCategoryItem(category_name):
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         category = session.query(Category).filter_by(name=category_name).one()
         newCategoryItem = CategoryItem(name=request.form['name'],
@@ -151,10 +183,12 @@ def newCategoryItem(category_name):
 @app.route('/category/<string:category_name>/<string:category_item_name>/edit',
            methods=['GET', 'POST'])
 def editCategoryItem(category_name, category_item_name):
+    if 'username' not in login_session:
+        return redirect('/login')
     category = session.query(Category).filter_by(name=category_name).one()
     categoryItem = session.query(CategoryItem).filter_by(
-                                                category_id=category.id,
-                                                name=category_item_name).one()
+        category_id=category.id,
+        name=category_item_name).one()
     if request.method == 'POST':
         categoryItem.name = request.form['name']
         categoryItem.description = request.form['description']
@@ -171,12 +205,14 @@ def editCategoryItem(category_name, category_item_name):
 @app.route('/category/<string:category_name>/<string:category_item_name>'
            '/delete', methods=['GET', 'POST'])
 def deleteCategoryItem(category_name, category_item_name):
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         print(category_item_name)
         category = session.query(Category).filter_by(name=category_name).one()
         categoryItem = session.query(CategoryItem).filter_by(
-                                                category_id=category.id,
-                                                name=category_item_name).one()
+            category_id=category.id,
+            name=category_item_name).one()
         session.delete(categoryItem)
         session.commit()
         return redirect(url_for('menu', category_name=category_name))
